+++ app/vmui/packages/vmui/src/components/Home/HomeLayout.tsx
@@ -21,7 +21,7 @@ const HomeLayout: FC = () => {
     <>
       <AppBar position="static">
         <Toolbar>
-          <Box mr={2} display="flex">
+          <Box display="flex">
             <Typography variant="h5">
               <span style={{fontWeight: "bolder"}}>VM</span>
               <span style={{fontWeight: "lighter"}}>UI</span>
@@ -43,24 +43,25 @@ const HomeLayout: FC = () => {
               Create an issue
             </Link>
           </div>
-          <Box flexGrow={1}>
+          <Box ml={4} flexGrow={1}>
             <ExecutionControls/>
           </Box>
           <DisplayTypeSwitch/>
           <UrlCopy url={fetchUrl}/>
         </Toolbar>
       </AppBar>
-      <Box display="flex" flexDirection="column" style={{height: "calc(100vh - 64px)"}}>
+      <Box display="flex" flexDirection="column" style={{minHeight: "calc(100vh - 64px)"}}>
         <Box m={2}>
           <QueryConfigurator/>
         </Box>
-        <Box flexShrink={1} style={{overflowY: "scroll"}}>
+        <Box flexShrink={1}>
           {isLoading && <Fade in={isLoading} style={{
             transitionDelay: isLoading ? "300ms" : "0ms",
           }}>
             <Box alignItems="center" flexDirection="column" display="flex"
               style={{
                 width: "100%",
+                maxWidth: "calc(100vh - 32px)",
                 position: "absolute",
                 height: "150px",
                 background: "linear-gradient(rgba(255,255,255,.7), rgba(255,255,255,.7), rgba(255,255,255,0))"