+++ app/vmui/packages/vmui/src/state/common/reducer.ts
@@ -2,6 +2,7 @@ import {DisplayType} from "../../components/Home/Configurator/DisplayTypeSwitch"
 import {TimeParams, TimePeriod} from "../../types";
 import {dateFromSeconds, getDurationFromPeriod, getTimeperiodForDuration} from "../../utils/time";
 import {getFromStorage} from "../../utils/storage";
+import {getDefaultServer} from "../../utils/default-server-url";
 
 export interface TimeState {
   duration: string;
@@ -15,6 +16,7 @@ export interface AppState {
   time: TimeState;
   queryControls: {
     autoRefresh: boolean;
+    autocomplete: boolean
   }
 }
 
@@ -28,9 +30,11 @@ export type Action =
     | { type: "RUN_QUERY"}
     | { type: "RUN_QUERY_TO_NOW"}
     | { type: "TOGGLE_AUTOREFRESH"}
+    | { type: "TOGGLE_AUTOCOMPLETE"}
+
 
 export const initialState: AppState = {
-  serverUrl: getFromStorage("PREFERRED_URL") as string || "https://", // https://demo.promlabs.com or https://play.victoriametrics.com/select/accounting/1/6a716b0f-38bc-4856-90ce-448fd713e3fe/prometheus",
+  serverUrl: getFromStorage("PREFERRED_URL") as string || getDefaultServer(), // https://demo.promlabs.com or https://play.victoriametrics.com/select/accounting/1/6a716b0f-38bc-4856-90ce-448fd713e3fe/prometheus",
   displayType: "chart",
   query: getFromStorage("LAST_QUERY") as string || "\n", // demo_memory_usage_bytes
   time: {
@@ -38,7 +42,8 @@ export const initialState: AppState = {
     period: getTimeperiodForDuration("1h")
   },
   queryControls: {
-    autoRefresh: false
+    autoRefresh: false,
+    autocomplete: getFromStorage("AUTOCOMPLETE") as boolean || false
   }
 };
 
@@ -97,6 +102,14 @@ export function reducer(state: AppState, action: Action): AppState {
         queryControls: {
           ...state.queryControls,
           autoRefresh: !state.queryControls.autoRefresh
+        }
+      };
+    case "TOGGLE_AUTOCOMPLETE":
+      return {
+        ...state,
+        queryControls: {
+          ...state.queryControls,
+          autocomplete: !state.queryControls.autocomplete
         }
       };
     case "RUN_QUERY":