+++ app/vmui/packages/vmui/src/utils/default-server-url.ts
@@ -1,0 +1,6 @@
+export const getDefaultServer = (): string => {
+  const {href} = window.location;
+  const regexp = /^http.+\/vmui/;
+  const [result] = href.match(regexp) || ["https://"];
+  return result.replace("vmui", "prometheus");
+};