+++ app/vmui/packages/vmui/src/utils/detect-os.ts
@@ -1,0 +1,13 @@
+const desktopOs = {
+  windows: "Windows",
+  mac: "Mac OS",
+  linux: "Linux"
+};
+
+export const getOs = () : string => {
+  return Object.values(desktopOs).find(os => navigator.userAgent.indexOf(os) >= 0) || "unknown";
+};
+
+export const isMacOs = (): boolean => {
+  return getOs() === desktopOs.mac;
+};