+++ lib/promscrape/discovery/gce/instance.go
@@ -93,6 +93,7 @@ type Instance struct {
 
 // NetworkInterface is network interface from https://cloud.google.com/compute/docs/reference/rest/v1/instances/list
 type NetworkInterface struct {
+	Name          string
 	Network       string
 	Subnetwork    string
 	NetworkIP     string
@@ -147,6 +148,10 @@ func (inst *Instance) appendTargetLabels(ms []map[string]string, project, tagSep
 		"__meta_gce_project":         project,
 		"__meta_gce_subnetwork":      iface.Subnetwork,
 		"__meta_gce_zone":            inst.Zone,
+	}
+	for _, iface := range inst.NetworkInterfaces {
+		ifaceName := discoveryutils.SanitizeLabelName(iface.Name)
+		m["__meta_gce_interface_ipv4_"+ifaceName] = iface.NetworkIP
 	}
 	if len(inst.Tags.Items) > 0 {
 		// We surround the separated list with the separator as well. This way regular expressions